refactor(profile-comics): added minor refactoring improvements
