+++ packages/frontend/src/components/profile/ProfileCollection/ProfileCollection.tsx
@@ -1,0 +1,22 @@
+import { useState } from 'react';
+import Actions from './Actions';
+
+import Comics from './Comics';
+
+const ProfileComics: React.FC = () => {
+  const [isEdit, setIsEdit] = useState(false);
+
+  const toggleEditHandler = () => {
+    setIsEdit((prevState) => !prevState);
+  };
+
+  return (
+    <section>
+      <h2 className="font-bold">Your Collection</h2>
+      <Comics isEdit={isEdit} />
+      <Actions onEdit={toggleEditHandler} />
+    </section>
+  );
+};
+
+export default ProfileComics;
